PHP Developer Assignment
"You work for a company that has tripled in size over the past few years and the way the vacation process works is no longer efficient, as it requires a combination of hand written applications, approvals, storage and maintenance. You are tasked with creating a portal where employees can request their vacation online, the manager receives a notification to approve or decline that request, and the information (time used, balances) are stored within the portal. "

database.php: represents the connection to the database through the PHP code
login.php: represents the login interface where the user is going to login and be transferred to the vacation_request page
vacation_request.php: represents the leave platform so as any employee can assign the days they are going to be away
manage_requests.php: in case the user who logged in is the manager then they are redirected to the manage_requests, which is only accessible to managers, so as they can approve or reject a leave request.